Fix headers in encoded output
The =ypart should not contain the part number, the =ybegin should when
there are more than 1 parts. In =yend the order of part and size was
different than the yenc2 test data from yenc.org.
The default line length was always printed in the =ybegin header
instead of the applied line length